feat/refactor: infraestructura de reportes, worker híbrido y optimización docker (#314)
* Agregando Archivos Base

- Se agregaron los DTO a utilizar
- Servicio inicial de traida de Datos

TODO: Revicion de Datos

* Atualizacion Endpoint para datos de exportación de reportes

Refer #277
- Se puede ver el sumary de datos
- se puede ver los datos paginados [aun en preubas]

* Testing : Quest Library para la Generacion de PDF

- Implementacion Basica de Quest PDF

* refactor(startup): separar configuración en partial classes

- Se marca la clase Startup como 'partial' para permitir su división.
- Extracción de servicios de autenticación e identidad a Startup.Auth.cs.
- Reorganización de la inyección de dependencias en Startup.Domain.cs.
- Limpieza de directivas 'using' no utilizadas en el archivo principal.

Este cambio no afecta la lógica de negocio, solo mejora la mantenibilidad
y legibilidad del punto de entrada de la aplicación.

Closed #302

* Feacture: Hanfire Service agregado al proyecto

- Se agrego hanfire worker al proyecto
- Se agregaron variables para controlar recurso como embebido o externo
- Se eliminaron dependencias de Quest

Closed #313

* refactor: optimización de dockerfiles y orquestación compose

- Implementación de multi-stage builds para reducir el footprint de las imágenes.
- Ajuste en SIGREF.API, AppHost y Frontend (Nginx).
- Optimización de archivos compose.yaml para la instancia central del servidor.
- Se eliminaron dependencias de compilación en las imágenes de runtime para mejorar la seguridad.
- Mejora en la persistencia de volúmenes y aislamiento de redes en Docker.

Closed #304

* feature: implementación de núcleo de reportes compartido

- Creada biblioteca SIGREF.Infrastructure.Reporting para centralizar QuestPDF.
- Configurada inyección de dependencias (DependencyInjection) para registro en API y Worker.
- Vinculado proyecto a SIGREF.API y SIGREF.Hangfire.Worker mediante referencias de proyecto.

Refer #309

* Chore: Ajustes menores referencias y variables

- ajustes en los logs
- ajustes de variables
- ajustes en las referencias

+++ BE/SIGREF.API/Dtos/Common/PagedResultDto.cs
@@ -7,7 +7,7 @@ public class PaginationDto
     public int CurrentPage { get; set; }
     public int PageSize { get; set; }
     // tipo ? para Keycloak
-    public int? TotalItems { get; set; }
+    public long? TotalItems { get; set; }
     // tipo ? para keycloak
     public int? TotalPages { get; set; }
     public bool HasPrevious { get; set; }
